change quiz repository name
