+++ app/src/main/java/com/space/quizapp/di/QuizModule.kt
@@ -1,7 +1,7 @@
 package com.space.quizapp.di
 
-import com.space.quizapp.data.repository.QuizRepositoryImpl
-import com.space.quizapp.domain.repository.QuizRepository
+import com.space.quizapp.data.repository.AvailableQuizRepositoryImpl
+import com.space.quizapp.domain.repository.AvailableQuizRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -15,5 +15,5 @@ abstract class QuizModule {
 
     @Binds
     @Singleton
-    abstract fun bindQuizRepository(quizRepositoryImpl: QuizRepositoryImpl): QuizRepository
+    abstract fun bindQuizRepository(quizRepositoryImpl: AvailableQuizRepositoryImpl): AvailableQuizRepository
 }
