+++ app/src/main/java/com/space/quizapp/domain/repository/AvailableQuizRepository.kt
@@ -1,0 +1,8 @@
+package com.space.quizapp.domain.repository
+
+import com.space.quizapp.domain.model.QuizModel
+import kotlinx.coroutines.flow.Flow
+
+interface AvailableQuizRepository {
+    suspend fun getAvailableQuizList(): Flow<List<QuizModel>>
+}
