--- app/src/main/java/com/space/quizapp/domain/repository/QuizRepository.kt
@@ -1,9 +1,0 @@
-package com.space.quizapp.domain.repository
-
-import com.space.quizapp.domain.model.QuizModel
-import kotlinx.coroutines.flow.Flow
-
-interface QuizRepository {
-    suspend fun getAvailableQuizList(): Flow<List<QuizModel>>
-
-}
