+++ app/src/main/java/com/space/quizapp/domain/usecase/quiz/AvailableQuizUseCase.kt
@@ -1,0 +1,10 @@
+package com.space.quizapp.domain.usecase.quiz
+
+import com.space.quizapp.domain.repository.AvailableQuizRepository
+import javax.inject.Inject
+
+class AvailableQuizUseCase @Inject constructor(
+    private val quizRepository: AvailableQuizRepository
+) {
+    suspend fun getAvailableQuiz() = quizRepository.getAvailableQuizList()
+}
